deployment/requirements.txt
streamlit
pandas
numpy
scikit-learn==1.6.1
Pillow
plotly
XGBoost